Add Dockerfile to project base directory
Add a dockerfile containing two targets to the base directory. The two
targets are `unit_testrunner` and `runtime`.

Building (with `docker build`) with argument `--target unit_testrunner`
will build an image which runs tests only on start up in a container.

Building with argument `--target runtime` will build an image which runs
the project on start up in a container.